pyproject.toml
[project]
name = "kelvin"
version = "0.1.0"
description = "Code examination tool"
readme = "README.md"
requires-python = "==3.12.*"

dependencies = [
    "beautifulsoup4==4.11.1",
    "bokeh==2.3.2",
    "cssselect==1.1.0",
    "django==4.2.16",
    "django-cas-ng==4.3.0",
    "django-notifications-hq==1.8.0",
    "django-redis>=5.4.0",
    "django-rq==2.10.2",
    "django-tasks-scheduler==1.3.4",
    "django-webpush==0.3.4",
    "imageio==2.6.1",
    "Jinja2==2.11.3",
    "lxml==5.1.0",
    "markdown==3.7",
    "MarkupSafe==2.0.1",
    "mdx-truly-sane-lists==1.3",
    "mosspy==1.0.9",
    "networkx==2.5",
    "numpy==1.26.4",
    "oauthlib==3.2.2",
    "pandas==2.2.1",
    "psycopg[binary]==3.2.1",
    "pygraphviz==1.7",
    "pymdown-extensions==10.10.1",
    "pyserde==0.25.1",
    "python-dotenv==1.0.1",
    "python-magic==0.4.27",
    "pyyaml==6.0.1",
    "requests==2.32.3",
    "requests_oauthlib==1.3.1",
    "six==1.16.0",
    "setuptools==69.1.0",
    "typing-extensions==4.11.0",
    "Unidecode==1.3.6",
    "uWSGI==2.0.24",
    "django-ipware==7.0.1",
    "openai>=2.6.1",
    "pycountry~=24.6.1",
    "django-ninja==1.4.5",
]

[tool.uv]
dev-dependencies = [
    "pre-commit>=3.8.0",
    "ruff==0.6.3",
    "pyrefly==0.18.0",
    "django-debug-toolbar>=6.0.0",
]

[tool.ruff]
target-version = "py312"
line-length = 100
indent-width = 4

include = [
    "api/**/*.py",
    "common/**/*.py",
    "evaluator/**/*.py",
    "kelvin/**/*.py",
    "survey/**/*.py",
    "web/**/*.py"
]
extend-exclude = [
    "**/migrations/**"
]
[tool.pyrefly]
project_includes = ["**/*"]
project_excludes = [
    "**/.[!/.]*",
    "**/*venv/**/*",
]
